[#1] feat: Input 컴포넌트 타입 확장
